Add helper to determine if UCS2 conversion should be performed
This will determine if a specific object type needs UCS2 character
encoding conversion performed or not at all. It also respects a
new global "ConvertUcs2ToUtf8" flag which allows to control the
behavior in general. For instance git does not require the encoding
conversion and works fine with UCS2 files so the flag can be set
to "False" in such a case.